Rename the library
